requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-docx==0.8.11
pytest==8.3.5
httpx==0.24.1
openai==1.12.0
python-dotenv==1.1.0
pytest-asyncio==0.26.0
pydantic-settings==2.1.0
tiktoken==0.5.2
aiofiles==23.2.1